requirements.txt
python-json-logger==2.0.7
PyYAML==6.0.1
python-dotenv==1.0.1